requirements.txt
pyasn1>=0.1.9
pyOpenSSL>=16.0.0
pycryptodome>=3.4
dateutils>=0.6.6
geopy>=1.11.0
idna>=2.1
profilehooks>=1.8.1
pycparser>=2.14
python-dateutil>=2.5.3
pytz>=2016.6.1
repoze.lru>=0.6
six>=1.10.0
xmltodict>=0.10.2
